refactor: add success & info to VariantIcon & add it to Alert
