+++ packages/Alert/package.json
@@ -32,7 +32,8 @@
   "dependencies": {
     "@welcome-ui/box": "^1.23.0",
     "@welcome-ui/system": "^1.20.0",
-    "@welcome-ui/text": "^1.20.0"
+    "@welcome-ui/text": "^1.20.0",
+    "@welcome-ui/variant-icon": "^1.27.0"
   },
   "peerDependencies": {
     "@xstyled/styled-components": "^1.17.0",
