+++ packages/Hint/doc.mdx
@@ -25,6 +25,8 @@ Said state information, use on Field.
   <Hint>default</Hint>
   <Hint variant="warning">warning</Hint>
   <Hint variant="error">error</Hint>
+  <Hint variant="info">info</Hint>
+  <Hint variant="success">success</Hint>
 </Playground>
 
 ## Properties
