+++ packages/Label/doc.mdx
@@ -37,11 +37,14 @@ With `required` property.
 
 ## Variants
 
-`error` or `warning` variant add specific badge.
+`error`, `warning`, `info` or `success` variant add specific badge, you can add a custom one as well.
 
 <Playground>
   <Label variant="warning">Warning variant</Label>
   <Label variant="error">Error variant</Label>
+  <Label variant="info">Info variant</Label>
+  <Label variant="success">Success variant</Label>
+  <Label icon="🐶">Custom icon</Label>
 </Playground>
 
 ## More content
