+++ packages/Toast/package.json
@@ -38,6 +38,7 @@
     "@welcome-ui/clear-button": "^1.29.0",
     "@welcome-ui/system": "^1.20.0",
     "@welcome-ui/utils": "^1.27.0",
+    "@welcome-ui/variant-icon": "^1.27.0",
     "react-spring": "^8.0.27",
     "toasted-notes": "^3.2.0"
   },
